+++ Dto/TenderWatchlistDto.cs
@@ -8,7 +8,7 @@ namespace MeroBolee.Dto
 {
     public class TenderWatchlistDto
     {
-        private int tender_Id;
+        private long tender_Id;
         private Guid tender_Code;
         private string tender_Title;
         private int category_Id;
@@ -27,7 +27,7 @@ namespace MeroBolee.Dto
         private int status_id;
         private AdminStatusEntity status;
 
-        public int Tender_Id { get => tender_Id; set => tender_Id = value; }
+        public long Tender_Id { get => tender_Id; set => tender_Id = value; }
         public Guid Tender_Code { get => tender_Code; set => tender_Code = value; }
         public string Tender_Title { get => tender_Title; set => tender_Title = value; }
         public int Category_Id { get => category_Id; set => category_Id = value; }
