+++ Model/BidderRequestDocEntity.cs
@@ -11,22 +11,22 @@ namespace MeroBolee.Model
     [Table("mb_bidder_request_doc")]
     public class BidderRequestDocEntity
     {
-        private int doc_id;
-        private int request_id;
+        private long doc_id;
+        private long request_id;
         private BidderRequestEntity bidderRequest;
         private string doc_location;
 
         [JsonIgnore]
         [Key,DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Column("id")]
-        public int Doc_id { get => doc_id; set => doc_id = value; }
+        public long Doc_id { get => doc_id; set => doc_id = value; }
 
         [Column("bidder_request_doc")]
         public string Document{ get => doc_location; set => doc_location = value; }
 
         [Column("request_id")]
         [ForeignKey("BidderRequest")]
-        public int Request_id { get => request_id; set => request_id = value; }
+        public long Request_id { get => request_id; set => request_id = value; }
 
         [JsonIgnore]
         public BidderRequestEntity BidderRequest { get => bidderRequest; set => bidderRequest = value; }
