+++ Model/WatchListEntity.cs
@@ -12,22 +12,30 @@ namespace MeroBolee.Model
     public class WatchListEntity
     {
         private int id;
-        private int user_Id;
+        private long user_Id;
         private UserEntity userEntity;
-        private int tender_Id;
+        private long tender_Id;
         private TenderEntity tenderEntity;
 
         [Key,DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Column("id")]
         public int Id { get => id; set => id = value; }
+
+
         [Column("user_id")]
         [ForeignKey("UserEntity")]       
-        public int User_Id { get => user_Id; set => user_Id = value; }
+        public long User_Id { get => user_Id; set => user_Id = value; }
+
+
         [JsonIgnore]
         public UserEntity UserEntity { get => userEntity; set => userEntity = value; }
+
+
         [Column("tender_id")]
         [ForeignKey("TenderEntity")]
-        public int Tender_Id { get => tender_Id; set => tender_Id = value; }
+        public long Tender_Id { get => tender_Id; set => tender_Id = value; }
+
+
         [JsonIgnore]
         public TenderEntity TenderEntity { get => tenderEntity; set => tenderEntity = value; }
     }
